LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_attributes_mouse-right-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: Buttons attribute on pointerover expected 0 but got 4
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: Buttons attribute on pointerenter expected 0 but got 4
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: Buttons attribute on pointermove expected 0 but got 4
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerdown event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerover event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerenter event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerdown event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerup event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerout event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerleave event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointermove event Reached unreachable code

done

Harness Error (FAIL), message = Unhandled rejection: assert_unreached: missing pointermove event Reached unreachable code

FAIL Test pointer events in the main document assert_equals: Buttons attribute on pointerover expected 0 but got 4
FAIL Test pointer events in an iframe assert_unreached: missing pointerover event Reached unreachable code
PASS mouse pointerup event is a PointerEvent event
PASS mouse pointerup.isTrusted value is true
PASS mouse pointerup.composed value is valid
PASS mouse pointerup.bubbles value is valid
PASS mouse pointerup.cancelable value is valid
PASS mouse pointerup.pressure value is valid
PASS mouse pointerup properties for pointerType = mouse
PASS mouse pointerup properties for pointerup
PASS mouse pointerout event is a PointerEvent event
PASS mouse pointerout.isTrusted value is true
PASS mouse pointerout.composed value is valid
PASS mouse pointerout.bubbles value is valid
PASS mouse pointerout.cancelable value is valid
PASS mouse pointerout.pressure value is valid
PASS mouse pointerout properties for pointerType = mouse
PASS mouse pointerleave event is a PointerEvent event
PASS mouse pointerleave.isTrusted value is true
PASS mouse pointerleave.composed value is valid
PASS mouse pointerleave.bubbles value is valid
PASS mouse pointerleave.cancelable value is valid
PASS mouse pointerleave.pressure value is valid
PASS mouse pointerleave properties for pointerType = mouse
